requirements.txt
streamlit==1.30.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
xgboost==2.0.3
scikit-learn==1.5.1
catboost==1.2.5
matplotlib==3.9.2
shap==0.45.1
lightgbm==4.3.0
pillow==10.4.0